+++ include/tvm/runtime/c_runtime_api.h
@@ -467,7 +467,7 @@ TVM_DLL int TVMArrayToDLPack(TVMArrayHandle from,
 
 /*!
  * \brief Delete (free) a DLManagedTensor's data.
- * \param dltensor Pointer to the DLManagedTensor. 
+ * \param dltensor Pointer to the DLManagedTensor.
  */
 TVM_DLL void TVMDLManagedTensorCallDeleter(DLManagedTensor* dltensor);
